1.renam. 2.标签回收成功的话,标签本身也会回收.

+++ src/main/java/com/sobey/cmop/mvc/service/iaas/MonitorElbServcie.java
@@ -40,7 +40,9 @@ public class MonitorElbServcie extends BaseSevcie {
 
 	@Transactional(readOnly = false)
 	public void deleteMonitorElb(Integer id) {
-		monitorElbDao.delete(id);
+		if (this.getMonitorElb(id) != null) {
+			monitorElbDao.delete(id);
+		}
 	}
 
 	/**
